+++ config/settings.py
@@ -69,8 +69,8 @@ class Config:
 
     # File Upload Configuration
     MAX_CONTENT_LENGTH = int(
-        os.environ.get("MAX_CONTENT_LENGTH", 5 * 1024 * 1024 * 1024)
-    )  # Default 2GB
+        os.environ.get("MAX_CONTENT_LENGTH", 10 * 1024 * 1024 * 1024)
+    )  # Default 10GB
     UPLOAD_FOLDER = os.environ.get("UPLOAD_FOLDER") or "uploads"
     ALLOWED_VIDEO_EXTENSIONS = {
         "mp4",